Separate gen3 plugin from default impl
