+++ .secrets.baseline
@@ -142,7 +142,7 @@
         "filename": "plugin-go/auth_impl_test.go",
         "hashed_secret": "00cafd126182e8a9e7c01bb2f0dfd00496be724f",
         "is_verified": false,
-        "line_number": 30
+        "line_number": 32
       }
     ],
     "plugin-python/plugin_test.py": [
@@ -164,5 +164,5 @@
       }
     ]
   },
-  "generated_at": "2025-04-17T19:17:59Z"
+  "generated_at": "2025-05-22T22:43:38Z"
 }
