+++ plugin-go/auth_impl_test.go
@@ -1,3 +1,5 @@
+//go:build ignore
+
 package main
 
 import (
